status produksi

+++ app/Http/Controllers/ProduksiBarangController.php
@@ -47,16 +47,29 @@ class ProduksiBarangController extends Controller
             })
 
             ->addColumn('aksi', function ($produksibarang) {
-                return '
-            <div class="">
-                <button onclick="editForm(`' . route('produksi.update', $produksibarang->id_produksi) . '`)" class="btn btn-xs btn-info btn-flat"><i class="fa fa-pencil"></i></button>
+                $html = '<div class="">';
+            if (
+                is_null($produksibarang->id_produksi) &&
+                    $produksibarang->status == StatusProduksiEnum::Terima
+                ){
+
+                $html .= '<button onclick="terimaProduksiBarang(`' . route('produksi.terima_produksi', $produksibarang->id_produksi) . '`)" class="btn btn-xs btn-primary btn-flat"><i class="fa fa-check"></i></button>
+
+                <button onclick="tolakProduksiBarang"(`' .route('produksi.tolak_produksi', $produksibarang->id_produksi) . '`)" class="btn btn-xs btn-danger btn-flat"><i class="fa fa-ban"></i></button>';
+            } else {
+                $html .=
+                '<button onclick="editForm(`' . route('produksi.update', $produksibarang->id_produksi) . '`)" class="btn btn-xs btn-info btn-flat"><i class="fa fa-pencil"></i></button>
+
                 <button onclick="deleteData(`' . route('produksi.destroy', $produksibarang->id_produksi) . '`)" class="btn btn-xs btn-danger btn-flat"><i class="fa fa-trash"></i></button>
-                <a href=' . route('detailProduksi.index', $produksibarang->id_produksi) . ' class="btn btn-xs btn-primary btn-flat">detail produksi</a>
-            </div>
-            ';
-            })
-            ->rawColumns(['aksi', 'jumlah', 'kode_produksi'])
-            ->make(true);
+
+                <a href=' . route('detailProduksi.index', $produksibarang->id_produksi) . ' class="btn btn-xs btn-primary btn-flat">detail produksi</a>';
+
+            }
+            $html .= '</div>';
+            return $html;
+        })
+        ->rawColumns(['aksi', 'jumlah', 'kode_produksi'])
+        ->make(true);
     }
 
     public function prosesProduksi(Request $request)
